Use StructType::get() to avoid creating new types
The StructType::create() used by the old implementation creates a new
unique struct type everytime it is called. This makes
Module::getOrInsertFunction() thinks the function types are mismatched
and adds an extra bitcast, while the types are really the same.
This could also cause assertion failures at some places because with the
constexpr bitcast the Value is no longer a Funtion despite having a
FunctionType.

Replace StructType::create() with StructType::get().

Fixes: 30c0b679f459 ("entry code insertion v0")
Signed-off-by: Jinghao Jia <jinghao7@illinois.edu>